+++ app/models/server/models/Settings.js
@@ -58,7 +58,7 @@ export class Settings extends Base {
 			filter._id = { $in: ids };
 		}
 
-		return this.find(filter, { fields: { _id: 1, value: 1, editor: 1 } });
+		return this.find(filter, { fields: { _id: 1, value: 1, editor: 1, enterprise: 1, invalidValue: 1, modules: 1 } });
 	}
 
 	findNotHiddenPublicUpdatedAfter(updatedAt) {
@@ -70,7 +70,7 @@ export class Settings extends Base {
 			},
 		};
 
-		return this.find(filter, { fields: { _id: 1, value: 1, editor: 1 } });
+		return this.find(filter, { fields: { _id: 1, value: 1, editor: 1, enterprise: 1, invalidValue: 1, modules: 1 } });
 	}
 
 	findNotHiddenPrivate() {
