+++ app/settings/client/lib/settings.ts
@@ -2,10 +2,7 @@ import { Meteor } from 'meteor/meteor';
 import { ReactiveDict } from 'meteor/reactive-dict';
 
 import { PublicSettingsCachedCollection } from '../../../../client/lib/settings/PublicSettingsCachedCollection';
-import { SettingsBase, SettingValue, ISettingRecord } from '../../lib/settings';
-import { hasLicense } from '../../../../ee/app/license/client';
-
-let isEnterprise = false;
+import { SettingsBase, SettingValue } from '../../lib/settings';
 
 class Settings extends SettingsBase {
 	cachedCollection = PublicSettingsCachedCollection.get()
@@ -18,43 +15,18 @@ class Settings extends SettingsBase {
 		return this.dict.get(_id);
 	}
 
-	getRecordValue(record: ISettingRecord): SettingValue {
-		const { value, invalidValue } = record;
-
-		if (!record.enterprise) {
-			return value;
-		}
-
-		if (!isEnterprise) {
-			return invalidValue;
-		}
-
-		if (!record.modules?.length) {
-			return value;
-		}
-
-		for (const moduleName of record.modules) {
-			if (!hasLicense(moduleName)) {
-				return invalidValue;
-			}
-		}
-
-		return value;
-	}
-
-	private _storeSettingValue(record: ISettingRecord, initialLoad: boolean): void {
-		const value = this.getRecordValue(record);
-		Meteor.settings[record._id] = value;
-		this.dict.set(record._id, value);
-		this.load(record._id, value, initialLoad);
+	private _storeSettingValue(record: { _id: string; value: SettingValue }, initialLoad: boolean): void {
+		Meteor.settings[record._id] = record.value;
+		this.dict.set(record._id, record.value);
+		this.load(record._id, record.value, initialLoad);
 	}
 
 	init(): void {
 		let initialLoad = true;
 		this.collection.find().observe({
-			added: (record: ISettingRecord) => this._storeSettingValue(record, initialLoad),
-			changed: (record: ISettingRecord) => this._storeSettingValue(record, initialLoad),
-			removed: (record: ISettingRecord) => {
+			added: (record: { _id: string; value: SettingValue }) => this._storeSettingValue(record, initialLoad),
+			changed: (record: { _id: string; value: SettingValue }) => this._storeSettingValue(record, initialLoad),
+			removed: (record: { _id: string }) => {
 				delete Meteor.settings[record._id];
 				this.dict.set(record._id, null);
 				this.load(record._id, undefined, initialLoad);
@@ -67,15 +39,3 @@ class Settings extends SettingsBase {
 export const settings = new Settings();
 
 settings.init();
-
-Meteor.startup(() => {
-	Meteor.call('license:isEnterprise', (err: any, result: any) => {
-		if (err) {
-			throw err;
-		}
-
-		if (result) {
-			isEnterprise = true;
-		}
-	});
-});
