+++ ee/client/index.js
@@ -4,4 +4,3 @@ import '../app/canned-responses/client/index';
 import '../app/engagement-dashboard/client/index';
 import '../app/license/client/index';
 import '../app/livechat-enterprise/client/index';
-import '../app/settings/client/index';
